% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccsr_dx.R
\name{ccsr_dx}
\alias{ccsr_dx}
\title{Identify CCSR categories for ICD-10-CM codes}
\usage{
ccsr_dx(icd_10_dx)
}
\arguments{
\item{icd_10_dx}{A vector of ICD-10-CM diagnosis codes}
}
\value{
A list charachter vectors with a of \code{length(icd_10_dx)}.
}
\description{
Given a vector of ICD-10-CM diagnosis codes, this function returns
a list of the Clinical Classifications Software Refined (CCSR)
categories associated with these codes
}
\details{
The Clinical Classifications Software Refined (CCSR) software
provides both one-to-one mapping (see \code{\link{classify_ccsr_dx1}})
and one-to-many (see \code{\link{ccsr_dx}}) mapping of ICD-10-CM
diagnosis codes to CCSR categories. The one-to-many mapping is
necessary because many ICD-10 codes span multiple meaningful
categories, and the identification of all conditions related to
a diagnosis code would be lost by categorizing some ICD-10-CM
codes into a single category.
\subsection{One-to-many}{

For example, consider the code \strong{I11.0} (\emph{Hypertensive heart
disease with heart failure}) which encompasses both \strong{heart
failure} (CCSR category CIR019) and \strong{hypertension with
complications} (CIR008). Classifying this code as heart
failure alone would miss meaningful information, but on the
other hand, some analyses require mutually exclusive
categorization.

This function addresses identifying all CCSR categories associated
with an ICD code. For example, consider a data.frame with an ICD-10
code per row:

\tabular{lc}{
\strong{pt_id} \tab \strong{\code{ICD10dx}}  \cr
A         \tab \code{K432}         \cr
A         \tab \code{A401}         \cr
B         \tab ...
}

The corresponding CCSR codes for these two codes are below. Note that
\code{K432} only has one CCSR category, while \code{A401} has two.

\tabular{lccc}{
\strong{\code{ICD10}}   \tab \strong{CCSR1} \tab \strong{CCSR2} \tab \strong{CCSR}(n)  \cr
\code{K432}      \tab   DIG010  \tab     ---   \tab   ...      \cr
\code{A401}      \tab   INF002  \tab   INF003  \tab   ...      \cr
...         \tab    ...    \tab     ...   \tab   ...
}

Running this function for patient A's two codes, results in \strong{three}
rows because \code{A401} has two categories (plus the single row for \code{K432}).
See \href{https://r4ds.had.co.nz/tidy-data.html}{the tidy data section}
of R for Data Science or the \href{https://doi.org/10.18637/jss.v059.i10}{corresponding paper}
for more on this conceptual approach.

\tabular{lcc}{
\strong{pt_id} \tab \strong{\code{ICD10dx}} \tab \strong{CCSR} \cr
A         \tab   \code{K432}      \tab   DIG010 \cr
A         \tab   \code{A401}      \tab   INF002 \cr
A         \tab   \code{A401}      \tab   INF003 \cr
B         \tab     ...       \tab    ...
}
}

\subsection{CCSR vs CCS}{

There are numerous differences between CCSR and CCS (the predecessor to CCSR).
The CCS classifies codes into multi-level categories in a hierarchical manner,
which allows users of CCS to use varying levels of specificity in their
classification (see \code{\link{classify_ccs}}), while the CCSR does \emph{not} have
multiple classification levles. Additionally, CCSR does not classify codes into
mutually exclusive categories (for ICD-10 diagnosis codes) and the categories
used in CCSR aren't the same as the old categories used in CCS.

See Appendix A of the \href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/DXCCSR-User-Guide-v2022-1.pdf}{CCSR user guide}
for more details on the differences between CCSR and CCS.
}
}
\examples{
library(dplyr)
library(tibble) # for tribble fxn
library(tidyr)  # for unnest

## Using a single ICD code
ccsr_dx("K432")
ccsr_dx("A401")

## When vectorized, returns list
ccsr_dx(c("K432", "A401")) \%>\% str()

## Use unnest to make it tidy
tibble::tribble(
  ~pt_id,  ~ICD10,
     "A",  "K432",
     "A",  "A401") \%>\%

  mutate(CCSR = ccsr_dx(ICD10)) \%>\%
  unnest_longer(CCSR)

}
\seealso{
\code{\link{classify_ccsr_dx1}} for identifying a single CCSR
category based on the principal diagnosis

\code{\link{classify_ccs}} for the legacy CCS categories
}
