% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_icd_format}
\alias{check_icd_format}
\title{Check that the ICD codes are in the correct format}
\usage{
check_icd_format(icd_codes)
}
\arguments{
\item{icd_codes}{Vector of ICD codes that should NOT have decimals}
}
\description{
Internal function to warn users if their ICD codes aren't in the
right format (i.e. have decimals, not character)
}
\keyword{internal}
