% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_CCS.R
\name{classify_ccs}
\alias{classify_ccs}
\title{Classify codes into CCS categories}
\source{
HCUP page for \href{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/ccs.jsp}{CCS with ICD-9}

HCUP page for \href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/ccsr_archive.jsp}{CCS beta with ICD-10}
}
\usage{
classify_ccs(icd_codes, code_type, level = NULL)
}
\arguments{
\item{icd_codes}{Vector of ICD codes, without decimals}

\item{code_type}{One of \verb{"dx9", "dx10", "pr9", or "pr10"}, corresponding
to the type and version of ICD codes used}

\item{level}{The level of CCS to use. The default is to use the
single-level CCS category}
}
\value{

}
\description{
Converts a vector of ICD diagnosis codes into a vector of Clinical
Classifications Software (CCS) categories using the HCUP's software
}
\details{
\subsection{Single vs Multi-level CCS categories}{

Unlike the CCS (CCS Refined), the original CCS classification
has "single-level" and "multi-level" categories.

This system classifies all diagnoses and procedures into unique groups.
The single-level CCS aggregates diagnoses into 285 mutually exclusive
categories and procedures into 231 mutually exclusive categories.

The multi-level CCS expands the single-level CCS into a hierarchical
system. The multi-level system has four levels for diagnoses and three
levels for procedures, which provide the opportunity to examine general
groupings or to assess very specific conditions and procedures. The
multi-level CCS groups single-level CCS categories into broader
body systems or condition categories (e.g., "Diseases of the Circulatory
System," "Mental Disorders," and "Injury").

An example using CCS for diagnosis codes is shown below:

\tabular{lll}{
\strong{CCS_lvl1} \tab \strong{CCS_level2} \tab \strong{CCS (single-level)} \cr
DX-1       \tab   DX-1.1       \tab \verb{DX1, DX2, DX3, DX9}   \cr
DX-1       \tab   DX-1.2       \tab \code{DX4}                  \cr
DX-1       \tab   DX-1.3       \tab \verb{DX5, DX6, DX7}        \cr
DX-1       \tab   DX-1.4       \tab \code{DX8}                  \cr
DX-1       \tab   DX-1.5       \tab \code{DX10}
}

It also splits single-level CCS categories to provide more detail.
For example, the single-level CCS diagnosis category \code{4} (\emph{Mycoses})
can be further split into 1.2.1 (\emph{Candidiasis of the mouth}) and
1.2.2 (\emph{Other mycoses}).

Further details can be found on the
\href{https://www.hcup-us.ahrq.gov/toolssoftware/ccs/ccsfactsheet.jsp}{CCS fact sheet}.
}

\subsection{Problems with the original notation}{

The notation used in the original CCS categories has a few limitations
in how it names categories.

First, the CCS category \code{CCS = '3'} maps to "\emph{Other bacterial
infections}" for diagnostic codes, but the same category (\code{CCS = '3'})
maps to "Laminectomy" for procedures. Second, the CCS category is
supposed to be treated as a string (because HCUP designs their software
for SAS), but R will appropriately assume these categories are numbers.

The third issue is the ambiguity of \emph{single-level} and \emph{multi-level} CCS
categories. In the original software, the first level of the \emph{multi-level}
CCS uses the same syntax as the \emph{single-level} categories. For example,
"4" represents "Mycoses" as a \emph{single-level} category, but maps to
"Diseases of the blood and blood-forming organs" as a \emph{multi-level} category!

This all turns out to be incredibly confusing as the same number \code{"3"} could
represent:
\itemize{
\item \emph{"Other bacterial infections"} if it's the single-level category
for a diagnosis
\item \emph{"Endocrine; nutritional; and metabolic diseases and immunity disorders"} if
it's the multi-level category for a diagnosis
\item \emph{"Laminectomy"} if it's the single-level category for a procedure
\item \emph{"Operations on the eye"} if it's the multi-level category for a procedure
}
}

\subsection{Notation used in this package}{

To address these issues, this package prepends \strong{"DX"}
or \strong{"PR"} before the \emph{default CCS category} (e.g. \code{3} becomes \code{DX3} or \code{PR3}
for diagnoses or procedures, respectively). For the \emph{multi-level categories},
the prefixes are \strong{"DX-"} and \strong{"PR-"}.

Although this is a trivial change for most applications, it is mentioned
here because (for the purposes of reproducibility) this notation should
be changed back to the original format for any publications or uses beyond this
package.
}
}
\seealso{
See \code{\link{explain_ccs}} to convert the CCS category into a
description of the category

\code{\link[hcup.data:CCS_dx9_map]{CCS_dx9_map}},
\code{\link[hcup.data:CCS_pr9_map]{CCS_pr9_map}},
\code{\link[hcup.data:CCS_dx10_map]{CCS_dx10_map}}, and
\code{\link[hcup.data:CCS_pr10_map]{CCS_pr10_map}} in the
\code{hcup.data} package for the datasets
}
