% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccsr_dx.R
\name{classify_ccsr_dx1}
\alias{classify_ccsr_dx1}
\title{Classify principal diagnosis into CCSR category}
\usage{
classify_ccsr_dx1(icd_10_dx1, setting = NULL)
}
\arguments{
\item{icd_10_dx1}{A vector of \strong{principal} ICD diagnosis codes
(without decimals)}

\item{setting}{Either "\code{inpatient}", (default) or "\code{outpatient}"
Note that this single default CCSR category only applies to
the principal diagnosis (for \code{inpatient}) and first-listed
diagnosis (for \code{outpatient}). These can also be abbreviated
as "ip" or "op".}
}
\value{
A vector of the default CCSR categories
}
\description{
Converts a vector of ICD-10 \strong{principal} diagnosis codes to the
default Clinical Classifications Software Refined (CCSR) category
based on the inpatient or outpatient setting
}
\details{
The Clinical Classifications Software Refined (CCSR) software
provides both one-to-one mapping (see \code{\link{classify_ccsr_dx1}})
and one-to-many (see \code{\link{ccsr_dx}}) mapping of ICD-10-CM
diagnosis codes to CCSR categories. The one-to-many mapping is
necessary because many ICD-10 codes span multiple meaningful
categories, and the identification of all conditions related to
a diagnosis code would be lost by categorizing some ICD-10-CM
codes into a single category.

For example, consider the code \strong{I11.0} (\emph{Hypertensive heart
disease with heart failure}) which encompasses both \strong{heart
failure} (CCSR category CIR019) and \strong{hypertension with
complications} (CIR008). Classifying this code as heart
failure alone would miss meaningful information, but on the
other hand, some analyses require mutually exclusive
categorization.

CCSR accomplishes this by providing a "default" CCSR category
based on the principal diagnosis (for inpatient data) and the
first-listed diagnosis (for outpatient data); see page 16 of
\href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/DXCCSR-User-Guide-v2022-1.pdf}{the CCSR DX user guide} for details.

This function addresses identifying the default CCSR category
based on the principal/first-listed diagnosis. For applications
utilizing all ICD-10-CM codes, such as risk adjustment and
identification of comorbidities, please use the
\code{\link{ccsr_dx}} function.
}
\examples{
classify_ccsr_dx1("I110", setting = "inpatient")

library(dplyr)
dplyr::tibble(ICD10 = c("G4730", "L563", "A001")) \%>\%
  mutate(CCSR_ip = classify_ccsr_dx1(ICD10, setting="inpatient"),
         CCSR_op = classify_ccsr_dx1(ICD10, setting="outpatient"))
}
\seealso{
\code{\link{ccsr_dx}} for identifying all CCSR categories in a
non-mutually exclusive method
}
