% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ccsr_pr.R
\name{classify_ccsr_pr}
\alias{classify_ccsr_pr}
\title{Classify ICD-10-PCS procedures into CCSR a category}
\source{
HCUP page for \href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/prccsr.jsp}{CCSR for procedures}
}
\usage{
classify_ccsr_pr(icd_pr)
}
\arguments{
\item{icd_pr}{A vector of ICD-10-PCS procedure codes, without decimals}
}
\value{
A vector of CCSR categories for procedures
}
\description{
Converts a vector of ICD-10 procedure codes into a vector of
Clinical Classifications Software Refined categories
for procedures using the HCUP's software
}
\examples{
classify_ccsr_pr("0016070")

icd_10_pr <- c("04L04DZ", "009Y40Z", "0F7C7DZ", "GZ2ZZZZ")
classify_ccsr_pr(icd_10_pr)

# Vectorized using `mutate()`
library(dplyr)

dplyr::tibble(icd_10 = icd_10_pr) \%>\%
  mutate(CCSR = classify_ccsr_pr(icd_10),
         explained = explain_ccsr(CCSR))
}
\seealso{
See \code{\link{explain_ccsr}} to convert the CCSR category into a
description of the category

\code{\link[hcup.data:CCSR_PR_mapping]{CCSR_PR_mapping}} in the
\code{hcup.data} package for the dataset
}
