% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_chronic.R
\name{classify_chronic}
\alias{classify_chronic}
\title{Classify ICD diagnosis codes into Chronic Condition Indicator categories}
\source{
HCUP page for \href{https://www.hcup-us.ahrq.gov/toolssoftware/chronic/chronic.jsp}{CCI with ICD-9}

HCUP page for \href{https://www.hcup-us.ahrq.gov/toolssoftware/chronic_icd10/chronic_icd10.jsp}{CCI with ICD-10}
}
\usage{
classify_chronic(icd_dx, icd_version)
}
\arguments{
\item{icd_dx}{A vector of ICD diagnosis codes (without decimals)}

\item{icd_version}{The version of ICD codes to be used. Can be
either ICD-10 (\code{icd_version = 10}) or ICD-9 (\code{icd_version = 9})}
}
\value{
A vector of CCI classifications. The levels returned depend on
the version of ICD codes. See Details section
}
\description{
Converts a vector of ICD diagnosis codes into a vector of Chronic
Condition Indicator (CCI) categories using the HCUP's software
}
\details{
Starting in v2021.1 (beta version), the CCI tool for ICD-10-CM was expanded
to identify four types of conditions. Therefore, running the
function with \code{icd_version = 10} will categorize ICD-10-CM codes into one of
the four conditions below:
\itemize{
\item \strong{Acute}: Examples include aortic embolism, bacterial infection, pregnancy,
and an initial encounter for an injury
\item \strong{Chronic}: Examples include malignant cancer, diabetes, obesity,
hypertension, and many mental health conditions
\item \strong{Both}: Examples include persistent asthma with (acute) exacerbation,
acute on chronic heart failure, and kidney transplant rejection
\item \strong{Not Applicable}: Examples include external cause of morbidity codes,
injury sequela codes, and codes starting with the letter Z for screening
or observation
}

In previous version (using ICD-9-CM), CCI classified conditions as \strong{Chronic}
or \strong{NonChronic}, so caution should be used when appying the CCI on data
using both ICD-9 and ICD-10.
}
\examples{
## Take an ICD dx code and return the CCI
classify_chronic(icd_dx = "G4730", icd_version = 10)
classify_chronic(icd_dx = "56081", icd_version = 9)

# Vectorized version
icd_codes <- c("G4730", "L563", "M25151")
classify_chronic(icd_codes, icd_version = 10)

# This works well in the dplyr workflow
library(dplyr)
pt_data <- tibble(
  Patients = c("A",     "B",    "C"),
  ICD_10   = c("G4730", "L563", "M25151")
)

pt_data \%>\%
  mutate(CCI = classify_chronic(ICD_10, icd_version = 10))
}
\seealso{
\code{\link[hcup.data:CCI_icd9]{CCI_icd9}} and
\code{\link[hcup.data:CCI_icd10]{CCI_icd10}} in the
\code{hcup.data} package for the datasets
}
