% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_pr.R
\name{classify_pr}
\alias{classify_pr}
\title{Classify Procedure Classes}
\source{
HCUP page for
\href{https://www.hcup-us.ahrq.gov/toolssoftware/procedure/procedure.jsp}{Procedure
Classes with ICD-9}

HCUP page for
\href{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10/procedure_icd10.jsp}{Procedure
Classes Refined with ICD-10}
}
\usage{
classify_pr(icd_pr)
}
\arguments{
\item{icd_pr}{A vector of ICD-9-CM or ICD-10-PCS codes for procedures. You do
not need to specify if the codes are ICD-9 or ICD-10.}
}
\value{
A vector of Procedure Classes, falling into one of four categories
listed in the details section
}
\description{
Converts a vector of ICD procedure codes into a vector of Procedure Classes
(for ICD-9) or Procedure Classes Refined (for ICD-10) using the HCUP's
software
}
\details{
The HCUP classifies procedures into the following four categories:
\itemize{
\item \strong{Minor Diagnostic}: Nonoperating room procedures that are diagnostic
\item \strong{Minor Therapeutic}: Nonoperating room procedures that are therapeutic
\item \strong{Major Diagnostic}: Procedures that are considered operating room
procedures that are performed for diagnostic reasons
\item \strong{Major Therapeutic}: Procedures that are considered operating room
procedures that are performed for therapeutic reasons
}
}
\section{Troubleshooting}{

This has not been tested extensively with ICD-9 codes in
practice, so users experiencing difficulty may want to implement their own
solutions. Please see \code{?hcup.data::proc_class_icd9} for the data used here, and use
\code{data(proc_class_icd9)} to adjust to your needs.
}

\examples{
classify_pr("00993ZX")
classify_pr("0015")

# Using the dplyr workflow
library(dplyr)
pt_data <- tibble(
  Patients = c("A",       "B",       "C"),
  ICD_10   = c("0PS443Z", "3E05305", "F13ZHZZ")
)

pt_data \%>\%
  mutate(pr_class = classify_pr(ICD_10))
\dontrun{
# Set of wrapper functions can be helpful for filtering
is_diagnostic_pr("F13ZHZZ")
is_major_pr("0PS443Z")

pt_data \%>\% filter(is_therapeutic_pr(ICD_10))
}

}
\seealso{
\code{\link[hcup.data:proc_class_icd10]{proc_class_icd10}} and
\code{\link[hcup.data:proc_class_icd9]{proc_class_icd9}} in the
\code{hcup.data} package for the datasets
}
