% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain_ccs}
\alias{explain_ccs}
\title{Look up CCS codes descriptions}
\usage{
explain_ccs(ccs, detailed = FALSE)
}
\arguments{
\item{ccs}{Vector of CCS codes to be explained. Can be a mix of diagnosis
and/or procedure CCS groups.}

\item{detailed}{Single logical value which indicates if more details should
be returned. The default is \code{FALSE}, in which case only the description of
the CCS category is returned. If set to \code{TRUE}, it will return a list of
named character vectors}
}
\description{
Look up CCS codes descriptions
}
\examples{
# Explain the CCS for diagnosis categories
explain_ccs("DX37")

# The same function works for procedures
explain_ccsr("PR28")

# Vectorized using `mutate()`
library(dplyr)
dplyr::tibble(
  CCS = c("DX37", "PR28", "PR-5.6", "DX-9.6.3"),
  Type = c("DX", "PR", "PR", "DX")) \%>\%
mutate(explained = explain_ccs(CCS))
}
