% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain_ccsr}
\alias{explain_ccsr}
\title{Look up CCSR category descriptions}
\usage{
explain_ccsr(ccsr, detailed = FALSE)
}
\arguments{
\item{ccsr}{Vector of CCSR codes to be explained. Can be a mix of diagnosis
and/or procedure CCSR groups.}

\item{detailed}{Single logical value which indicates if more details should
be returned. The default is \code{FALSE}, in which case only the description of
the CCSR category is returned. If set to \code{TRUE}, it will return a list of
named character vectors}
}
\description{
Look up CCSR category descriptions
}
\examples{
# Explain the CCSR for diagnosis categories
explain_ccsr("SKN003")

# The same function works for procedures
explain_ccsr("LYM007")

# Vectorized using `mutate()`
library(dplyr)
dplyr::tibble(
  CCSR = c("NVS011", "DIG021", "MBD028", "URN001", "RES010", "CAR004"),
  Type = c("DX", "DX", "DX", "PR", "PR", "PR")) \%>\%
mutate(explained = explain_ccsr(CCSR))

## Examples using the detailed option
detailed <- tibble(CCSR_PR = c("URN001", "RES010", "CAR004")) \%>\%
  mutate(Explained = explain_ccsr(CCSR_PR),
         details   = explain_ccsr(CCSR_PR, detailed = TRUE))

library(purrr) # use purrr to extract one option
detailed \%>\%
  mutate(domain = map_chr(details, "clinical_domain"))

library(tidyr) ## use unnest_wider to see all details
detailed \%>\%
  unnest_wider(details)
}
