% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcup-package.R
\docType{package}
\name{hcup-package}
\alias{hcup}
\alias{hcup-package}
\title{hcup: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hunter Ratliff \email{hunterratliff.cv@gmail.com} (\href{https://orcid.org/0000-0003-4978-1998}{ORCID})

}
\keyword{internal}
