% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{heuristic_msg}
\alias{heuristic_msg}
\title{Provide a message if heuristic is used}
\usage{
heuristic_msg(var_name, default)
}
\arguments{
\item{var_name}{Variable that was not supplied}

\item{default}{The default value to be used}
}
\description{
Provide a message if heuristic is used
}
\keyword{internal}
