% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_decimals}
\alias{warn_decimals}
\title{Warn if decimals in ICD codes}
\usage{
warn_decimals(icd_codes)
}
\arguments{
\item{icd_codes}{Vector of ICD codes that should NOT have decimals}
}
\description{
Internal function to warn users if their ICD codes aren't in the
right format (i.e. have decimals)
}
\keyword{internal}
