% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcup-package.R
\docType{package}
\name{hcup-package}
\alias{hcup}
\alias{hcup-package}
\title{hcup: Implementation of HCUP Software Tools in R}
\description{
Adapts the software provided by the Healthcare Cost and Utilization Project (HCUP) for analyses of administrative database using International Classification of Diseases (ICD) codes within the R statistical environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hunterratliff1.github.io/hcup}
}

}
\author{
\strong{Maintainer}: Hunter Ratliff \email{hunterratliff.cv@gmail.com} (\href{https://orcid.org/0000-0003-4978-1998}{ORCID})

}
\keyword{internal}
